[#PRX-250 Possible race condition for receiving n-event-rsp]
the scheduler threads can delete the empty dir during the naction
response write (before which the files have already been deleted by the
handler) this could mean the directory doesn't exist anymore. this fix
is to check for this case before trying to delete.

+++ dcm4chee-proxy-service/src/main/java/org/dcm4chee/proxy/dimse/StgCmt.java
@@ -698,7 +698,7 @@ public class StgCmt extends AbstractDicomService {
                         asAccepted.writeDimseRSP(pc, cmd, rspData);
                         File path = new File(file.getParent());
                         //null check since scheduler can beat this thread to deleting the empty file
-                        if (path!=null && path.list().length == 0)
+                        if (path!=null && path.list() !=null && path.list().length == 0)
                             path.delete();
                     } catch (IOException e) {
                         LOG.error(asAccepted
